Rename to counter.dart
