+++ lib/sandbox/CounterWithRiverpod/CounterPage.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:origin_master_2024_flutter/sandbox/CounterWithRiverpod/Counter.dart';
+import 'package:origin_master_2024_flutter/sandbox/CounterWithRiverpod/counter.dart';
 
 class CounterPage extends HookConsumerWidget {
   const CounterPage({Key? key}) : super(key: key);
